requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.8
pydantic-settings>=2.5
python-dotenv>=1.0
httpx>=0.27
aiohttp>=3.9,<4
msal>=1.28

# Bot Framework SDK (conector clásico)
botbuilder-core==4.14.7
botbuilder-schema==4.14.7
botframework-connector==4.14.7
